+++ tests/Testing/Concerns/InteractsWithExternalRequestsTest.php
@@ -423,7 +423,7 @@ class InteractsWithExternalRequestsTest extends FrameworkTestCase {
 			fn () => new DateTime(),
 		);
 
-		$this->expectException( RuntimeException::class );
+		$this->expectException( InvalidArgumentException::class );
 		$this->expectExceptionMessage( 'Unknown response type returned for faked request to [https://example.com/]. Expected a (Mantle\Testing\Mock_Http_Response|Mantle\Contracts\Support\Arrayable|WP_Error|array), got object.' );
 
 		Http::get( 'https://example.com/' );
@@ -448,5 +448,47 @@ class InteractsWithExternalRequestsTest extends FrameworkTestCase {
 		$this->assertEquals( 'https://example.com/?example=here', $object->url() );
 		$this->assertEquals( Http_Method::GET->value, $object->method() );
 		$this->assertEquals( Http_Method::GET, $object->enum_method() );
+	}
+
+	public function test_fake_request_with_snapshot(): void {
+		$this->fake_request( '*alley.com/*' )->with_snapshot();
+		// $this->fake_request( 'https://alley.com/wp-json/*' )->with_snapshot();
+
+		$this->assertEquals( 404, Http::get( 'https://alley.com/not-found-ever' )->status() );
+
+		$response = Http::get( 'https://alley.com/wp-json/wp/v2/posts/5038' );
+
+		$this->assertEquals( 200, $response->status() );
+		$this->assertNotEmpty( $response->json() );
+
+		// Assert that the request was made but was not a "real" request.
+		$this->assertRequestSent();
+		$this->assertRequestSent( 'https://alley.com/not-found-ever' );
+		$this->assertRequestSent( 'https://alley.com/wp-json/wp/v2/posts/5038' );
+
+		$this->assertActualRequestNotSent();
+		$this->assertActualRequestNotSent( 'https://alley.com/not-found-ever' );
+		$this->assertActualRequestNotSent( 'https://alley.com/wp-json/wp/v2/posts/5038' );
+
+		$this->assertFileExists(
+			__DIR__ . '/__http_snapshots__/InteractsWithExternalRequestsTest/test_fake_request_with_snapshot-get-https-alley-com-not-found-ever.json',
+		);
+
+		$this->assertFileExists(
+			__DIR__ . '/__http_snapshots__/InteractsWithExternalRequestsTest/test_fake_request_with_snapshot-get-https-alley-com-wp-json-wp-v2-posts-5038.json',
+		);
+	}
+
+	public function test_fake_request_with_snapshot_named(): void {
+		$this->fake_request( 'http://alley.com/not-found-with-body' )->with_snapshot( 'snapshot-id' );
+
+		$this->assertEquals( 404, Http::post( 'http://alley.com/not-found-with-body', [
+			'example' => 'here',
+		] )->status() );
+
+		// Assert that the expected snapshot file exists.
+		$this->assertFileExists(
+			__DIR__ . '/__http_snapshots__/InteractsWithExternalRequestsTest/test_fake_request_with_snapshot_named-snapshot-id.json',
+		);
 	}
 }